requirements.txt
python-dotenv
requests
SQLAlchemy
pandas
streamlit